requirements.txt
torch>=2.4.0
numpy>=1.24.0
gunicorn
scipy>=1.14.1
scikit-learn>=0.24.0
flask>=2.0.0
pillow>=9.0.0
gdown>=4.7.3
